Add contact form / messages (#22)
Add a contact form that creates messages in the database. People visiting the website can create messages through the contact form. Messages are directed to the site admin and hence there is no need for a show or an index action. Possible next step: Send a daily overview of the messages via email.