add published and parent_event_id to events (#25)
This is some preparation work to allow external people to suggest new events and edits to events. If an edit is suggested, a new, unpublished event will be created, that is linked to the original event. If a new event is suggested, it will also be unpublished first, until an admin publishes the event.